Port InputStreamReaderTest
